// src/communications/calls/item/operations/item/commsOperationItemRequestBuilderDeleteRequestConfiguration.ts
import type {RequestOption} from '@microsoft/kiota-abstractions';

export interface CommsOperationItemRequestBuilderDeleteRequestConfiguration {
    /**
     * Request headers
     */
    headers?: Record<string, string[]> | undefined;
    /**
     * Request options
     */
    options?: RequestOption[] | undefined;
}