// src/communications/calls/item/operations/item/commsOperationItemRequestBuilderPatchRequestConfiguration.ts
import type {RequestOption} from '@microsoft/kiota-abstractions';

export interface CommsOperationItemRequestBuilderPatchRequestConfiguration {
    /**
     * Request headers
     */
    headers?: Record<string, string[]> | undefined;
    /**
     * Request options
     */
    options?: RequestOption[] | undefined;
}